fix files according to PR comments. Refactor store. Create selectors files
