--- src/app/shared/pipes/toArray.pipe.ts
@@ -1,11 +1,0 @@
-import { Pipe, PipeTransform } from '@angular/core';
-
-@Pipe({ name: 'toArray' })
-export class ToArrayPipe implements PipeTransform {
-  transform(value: number | null) {
-    if (value) {
-      return new Array(value / 6);
-    }
-    return value;
-  }
-}
