+++ src/app/store/product-type-store/product-type.selectors.ts
@@ -1,0 +1,11 @@
+import { AppStateInterface } from '../../shared/interfaces/interfaces';
+import { createSelector } from '@ngrx/store';
+import { AppState } from './product-type.reducer';
+
+export const productTypeFeature = (state: AppStateInterface) =>
+  state.productType;
+
+export const productTypeSelector = createSelector(
+  productTypeFeature,
+  (state: AppState) => state.productType
+);
